+++ src/components/Task/index.tsx
@@ -3,3 +3,4 @@ export * from './List';
 export * from './Item';
 export * from './ListContainerView';
 export * from './CustomCalendarHeader';
+export * from './BottomSheet';
