+++ src/schemes/task/api.ts
@@ -60,6 +60,14 @@ const addTaskRequestScheme = todoTaskScheme.omit({ id: true, status: true, taskC
   routineCondition: taskRoutineConditionScheme.nullable(),
 });
 
+const taskFormScheme = addTaskRequestScheme.omit({ routineCondition: true }).extend({
+  routineCondition: taskRoutineConditionScheme.omit({ startDate: true, endDate: true, cycle: true }).extend({
+    startDate: z.string().nullable().describe('루틴 시작일자'),
+    endDate: z.string().nullable().describe('루틴 종료일자'),
+    cycle: TASK_ROUTINE_CYCLE_ENUM.nullable().describe('루틴 주기'),
+  }),
+});
+
 const updateTodoTaskStatusResponseScheme = BaseResponseScheme.extend({
   data: z.number().describe('완료된 task id'),
 });
@@ -96,6 +104,7 @@ export {
   fetchTaskListResponseDataScheme,
   fetchTaskListResponseScheme,
   addTaskRequestScheme,
+  taskFormScheme,
   updateTodoTaskStatusResponseScheme,
   fetchDowithTaskRequestScheme,
   fetchDowithTaskResponseScheme,
